requirements.txt
python-telegram-bot==20.3
httpx==0.24.0
python-dotenv==1.0.0
kink==0.6.6
pytz==2023.3

pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.0.0
pytest-mock==3.10.0
pytest-freezegun==0.4.2
greenlet==2.0.2

